Added Team edit functions and Identity directives.

+++ src/main/java/com/jda/crystalball/friday/repositories/TeamRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.jda.crystalball.friday.entities.TeamEntity;
 
 @Repository
-public interface TeamRepository extends JpaRepository<TeamEntity, Long>{
+public interface TeamRepository extends JpaRepository<TeamEntity, Integer>{
 
 }
